requirements.txt
streamlit
python-dotenv
requests
Pillow
diagrams

